feat: big changes for candidate 2.0 release
Splits Sinister into a library, and two runtimes:

1. Pride:
Same as the old runtime, updated to match the new changes

2. Sloth:
Supports side loading the teamcode folder, like fast load, but fixes a a
lot of issues

This will require changes to the test runtime for Sinister, which may be
added as a third, so far unnamed runtime, which is then used by the test
runtime, rather than the current hack system.

These changes are very breaking.

Notes:

SinisterFilter renamed to Scanner and has a nearly totally new api.

Now @NoUnload exists

These changes are not final, still to do:
1. add DAG tooling to Util somewhere, allowing for filters to depend on
   each other nicely.
2. write the test runtime
3. investigate reworking this build system structure to prefer
   `includeBuild` as part of a Dairy-wide change
4. see about removing no longer needed dependencies where possible, and
   not exposing the util library as api